refactor: 사전 설문 페이지에서 SurveyHeader 공통 컴포넌트 분리
