+++ front/src/Components/Header/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { styled, css } from "styled-components";
-import OptionalHeader from "../Header/OptionalHeader/index";
+import OptionalHeader from "./OptionalHeader/index";
 import HyundaiLogo from "../../assets/icons/hyundai-logo.svg";
 import HyundaiWhiteLogo from "../../assets/icons/hyundai-logo-white.svg";
 import { useLocation } from "react-router-dom";
