+++ front/src/Components/Survey/SurveyOption/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { css, styled } from "styled-components";
-import RadioBtn from "../Common/RadioBtn";
-import Check24BlueIcon from "../../assets/checkcircle/check-24-blue.svg";
+import RadioBtn from "../../Common/RadioBtn";
+import Check24BlueIcon from "../../../assets/checkcircle/check-24-blue.svg";
 
 const SurveyOption = ({ label, index, value, name, selected, onChange }) => {
   return (
