+++ front/src/Pages/SurveyPage/AgeSurvey/index.jsx
@@ -1,9 +1,9 @@
 import * as S from "../styles";
 import React, { useState } from "react";
-import SurveyOption from "../../../Components/SurveyOption";
+import SurveyOption from "../../../Components/Survey/SurveyOption";
 import Button from "../../../Components/Common/Button/Button";
-import PageIndicator from "../../../Components/PageIndicator";
 import { css } from "styled-components";
+import SurveyHeader from "../../../Components/Survey/SurveyHeader";
 
 const ageOptions = [
   {
@@ -31,12 +31,7 @@ const AgeSurvey = ({ buttonHandler }) => {
   };
   return (
     <S.SurveyContent>
-      <S.SurveyHeader>
-        <S.SurveyTitle>
-          <strong>나이</strong>를 알려주세요.
-        </S.SurveyTitle>
-        <PageIndicator crntPage={1} totalPage={2} />
-      </S.SurveyHeader>
+      <SurveyHeader index={1} />
 
       <S.SurveyOptions>
         {ageOptions.map((option, index) => (
@@ -50,6 +45,7 @@ const AgeSurvey = ({ buttonHandler }) => {
           />
         ))}
       </S.SurveyOptions>
+
       <Button text="다음" style={surveyBtnStyle} onClick={buttonHandler} />
     </S.SurveyContent>
   );
