+++ front/src/Pages/SurveyPage/styles/index.jsx
@@ -6,18 +6,7 @@ export const SurveyOptions = styled.div`
   gap: 12px;
   margin-top: 20px;
 `;
-export const SurveyTitle = styled.span`
-  ${({ theme }) => theme.font.Extra1}
 
-  strong {
-    font-weight: bold;
-  }
-`;
-export const SurveyHeader = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`;
 export const SurveyContent = styled.div`
   width: 608px;
   height: 100%;
